# docs/swagger.yaml
definitions:
  controllers.CartInput:
    properties:
      productId:
        type: integer
      total:
        type: integer
    type: object
  controllers.CategoryInput:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  models.Cart:
    properties:
      createdAt:
        type: string
      date:
        type: string
      id:
        type: integer
      price:
        type: integer
      productId:
        type: integer
      total:
        type: integer
      updatedAt:
        type: string
      usersId:
        type: integer
    type: object
  models.Category:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
...
  models.Product:
    properties:
      ImageUrl:
        type: string
      categoryID:
        type: integer
      condition:
        type: string
      createdAT:
        type: string
      description:
        type: string
      heavy:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
      updatedAT:
        type: string
      userID:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /cart:
    get:
      description: Get list carts refrence By userID
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cart'
      security:
      - BearerToken: []
      summary: Get carts By User Id
      tags:
      - Cart
    post:
      description: Creating a new Cart.
      parameters:
      - description: the body to create a new Cart
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.CartInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cart'
      security:
      - BearerToken: []
      summary: Create New Cart.
      tags:
      - Cart
  /cart/{cartId}/order/{orderId}:
    get:
      description: move a product from cart to Order Detail.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: cartid
        in: path
        name: cartId
        required: true
        type: string
      - description: orderId
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: cart Product to order detail.
      tags:
      - Cart
  /cart/{id}:
    delete:
      description: Delete a Cart by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Cart id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Cart.
      tags:
      - Cart
    patch:
      description: Update Cart by id.
      parameters:
      - description: the body to update a new Cart
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.CartInput'
      - description: Cart id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cart'
      security:
      - BearerToken: []
      summary: Update Cart.
      tags:
      - Cart
  /categories:
    get:
      description: Get list Category refrence By userID
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      security:
      - BearerToken: []
      summary: Get Category By User Id
      tags:
      - Category
    post:
      description: Creating a new Category.
      parameters:
      - description: the body to create a new Category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.CategoryInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      security:
      - BearerToken: []
      summary: Create New Category.
      tags:
      - Category
  /categories/{id}:
    delete:
      description: Delete a category by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Category.
      tags:
      - Category
    patch:
      description: Update Category by id.
      parameters:
      - description: the body to update a new Category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.CategoryInput'
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      security:
      - BearerToken: []
      summary: Update Category.
      tags:
      - Category
...
swagger: "2.0"